+++ jklib/django/db/tests.py
@@ -20,11 +20,11 @@ class ModelTestCase(ImprovedTestCase):
         Assertions for field constraints (unique, required, choices, etc.)
     """
 
+    model_class = None
+
     # ----------------------------------------
     # Properties
     # ----------------------------------------
-    model_class = None
-
     @property
     def common_errors(self):
         """
@@ -36,25 +36,6 @@ class ModelTestCase(ImprovedTestCase):
     # ----------------------------------------
     # Assertions
     # ----------------------------------------
-    def assert_fields_are_required(self, valid_payload, fields=None):
-        """
-        Tests that the required fields are truly required
-        For each field, we will:
-            Use a valid payload
-            Remove only the specific field
-            Try to create the object
-        :param dict valid_payload: A valid payload for the service
-        :param [str] fields: List of fields to check. Defaults to self.required_fields
-        """
-        if fields is None:
-            fields = self.required_fields
-        for field in fields:
-            with transaction.atomic():
-                payload = valid_payload.copy()
-                payload[field] = None
-                with self.assertRaises(self.common_errors):
-                    self.model_class(**payload).save()
-
     def assert_instance_count_equals(self, n):
         """Tests the number of instances in the database for our model"""
         assert self.model_class.objects.count() == n